+++ Weilefood/UI/Market/ViewController/ProductInfoVC.m
@@ -351,7 +351,7 @@ static NSString *const kCellIdentifier = @"MYCELL";
             [InputQuantityVC inputQuantityWithEnterBlock:^(InputQuantityVC *inputQuantityVC, NSInteger quantity) {
                 _strong_check(self);
                 // TODO 加入购物车
-                DLog(@"%ld", quantity);
+                DLog(@"%ld", (long)quantity);
                 [inputQuantityVC dismissSelf];
             }];
         }];
@@ -372,7 +372,7 @@ static NSString *const kCellIdentifier = @"MYCELL";
                 [InputQuantityVC inputQuantityWithEnterBlock:^(InputQuantityVC *inputQuantityVC, NSInteger quantity) {
                     _strong_check(self);
                     // TODO 立即购买
-                    DLog(@"%ld", quantity);
+                    DLog(@"%ld", (long)quantity);
                     [inputQuantityVC dismissSelf];
                 }];
             }];
